Restructure project so utils are part of ecp library
