+++ eCP/app/main.cpp
@@ -1,30 +1,30 @@
 #include <iostream>
 #include <ittnotify.h>
 
-#include <eCP/eCP.hpp>
-#include <utility.hpp>
+#include <eCP/index/eCP.hpp>
+#include <eCP/utilities/utility.hpp>
 
 int main()
 {
     /* For vtune params */
-    const int L = 3;           // L parameter - number of levels in index
-    const int metric = 0;      // Distance metric - 0 = euclidean - 1 = angular
-    const int k = 100;         // number points to return
-    const int b = 40;          // number clusters to search
-    const int p = 150000;      // number of vectors
-    const int d = 128;         // dimensions of vector
-    const int r = 1000;        // upper bound of generated vectors
-    const int qs = 15000;      // queries to make on created index
+//    const int L = 3;           // L parameter - number of levels in index
+//    const int metric = 0;      // Distance metric - 0 = euclidean - 1 = angular
+//    const int k = 100;         // number points to return
+//    const int b = 40;          // number clusters to search
+//    const int p = 150000;      // number of vectors
+//    const int d = 128;         // dimensions of vector
+//    const int r = 1000;        // upper bound of generated vectors
+//    const int qs = 15000;      // queries to make on created index
 
     /* For debugging params */
-//     const int L = 3;            // L parameter - number of levels in index
-//     const int metric = 0;       // Distance metric - 0 = euclidean - 1 = angular
-//     const int k = 2;            // number points to return
-//     const int b = 2;            // number clusters to search
-//     const int p = 120000;           // number of vectors
-//     const int d = 128;          // dimensions of vector
-//     const int r = 1000;         // upper bound of generated vectors
-//     const int qs = 15;          // queries to make on created index
+     const int L = 3;            // L parameter - number of levels in index
+     const int metric = 0;       // Distance metric - 0 = euclidean - 1 = angular
+     const int k = 2;            // number points to return
+     const int b = 2;            // number clusters to search
+     const int p = 12000;           // number of vectors
+     const int d = 128;          // dimensions of vector
+     const int r = 1000;         // upper bound of generated vectors
+     const int qs = 15;          // queries to make on created index
 
     /* Setup ITTAPI instrumentation domain */
     __itt_domain *domain_build = __itt_domain_create("ECP.BENCHMARKING.BUILD");
@@ -49,8 +49,8 @@ int main()
     __itt_task_end(domain_query);
 
     /* Debugging */
-//    utility::print_clusters(index->top_level);
-//    utility::print_index_levels(index->top_level);
+    utility::print_clusters(index->top_level);
+    utility::print_index_levels(index->top_level);
 //    utility::print_query_results(result, q, k, index->dataset);
 
     /* Clean up */
