+++ eCP/include/eCP/index/query-processing.hpp
@@ -11,7 +11,7 @@
  */
 
 #include <vector>
-#include <eCP/data_structure.hpp>
+#include <eCP/index/data_structure.hpp>
 
 namespace query_processing 
 {
